+++ libs/locale/src/lib/locale-switcher/locale-switcher.stories.ts
@@ -3,9 +3,10 @@ import { applicationConfig, Meta, moduleMetadata } from '@storybook/angular';
 import DeLocaleModule, { DE_LOCALE } from '@spryker/locale/locales/de';
 import EnLocaleModule, { EN_LOCALE } from '@spryker/locale/locales/en';
 import TrLocaleModule, { TR_LOCALE } from '@spryker/locale/locales/tr';
+import ElLocaleModule, { EL_LOCALE } from '@spryker/locale/locales/el';
 import { LocaleModule } from '../locale.module';
 
-const locales = { English: EN_LOCALE, German: DE_LOCALE, Turkish: TR_LOCALE };
+const locales = { English: EN_LOCALE, German: DE_LOCALE, Turkish: TR_LOCALE, Greek: EL_LOCALE };
 
 export default {
     title: 'LocaleSwitcherComponent',
@@ -16,6 +17,7 @@ export default {
                 importProvidersFrom(EnLocaleModule),
                 importProvidersFrom(DeLocaleModule),
                 importProvidersFrom(TrLocaleModule),
+                importProvidersFrom(ElLocaleModule),
             ],
         }),
         moduleMetadata({