+++ libs/table.feature.settings/src/i18n/el.ts
@@ -1,0 +1,9 @@
+import { I18nLocaleDataPackage } from '@spryker/locale';
+
+export default {
+    name: 'table.feature.settings',
+    data: {
+        title: 'Ρυθμίσεις πίνακα',
+        reset: 'Επαναφορά',
+    },
+} as I18nLocaleDataPackage;