+++ libs/unsaved-changes/src/i18n/el.ts
@@ -1,0 +1,10 @@
+import { I18nLocaleDataPackage } from '@spryker/locale';
+
+export default {
+    name: 'unsaved-changes',
+    data: {
+        'confirmation-title': 'Απόρριψη μη αποθηκευμένων αλλαγών;',
+        'confirmation-ok': 'Απόρριψη',
+        'confirmation-cancel': 'Ακύρωση',
+    },
+} as I18nLocaleDataPackage;