+++ libs/table.column.button-action/CHANGELOG.md
@@ -3,7 +3,7 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [1.0.11-next.0](http://172.31.0.22:9292/spryker-internal-ci/ui-components/compare/@spryker/table.column.button-action@1.0.10...@spryker/table.column.button-action@1.0.11-next.0) (2025-09-12)
+## [1.0.11](http://172.31.0.22:9292/spryker-internal-ci/ui-components/compare/@spryker/table.column.button-action@1.0.10...@spryker/table.column.button-action@1.0.11) (2025-09-19)
 
 **Note:** Version bump only for package @spryker/table.column.button-action